Separate common tests
